+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Anticipated release: September 21st, 2020
 
 #### 🚀 New features
-
 - Updated inputs in the Estimated Quarterly Incentive Payments table to use aria-labelledby for better UX/a11y ([#2503])
+- Change order of fields and field/title names on Cost Allocation and Other Funding + Budget and FFP pages. ([#2322])
 
 #### 🐛 Bugs fixed
 
@@ -17,3 +17,4 @@ Anticipated release: September 21st, 2020
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2503]: https://github.com/CMSgov/eAPD/issues/2503
+[#2322]: https://github.com/CMSgov/eAPD/issues/2322